Création d'un nouveau projet versionné

+++ TP2_Bieres_COURTEL/src/tp2_bieres_courtel/BouteilleBiere.java
@@ -22,7 +22,7 @@ public class BouteilleBiere {
  
     }
     
-    public BouteilleBiere(String unNom, float unDegre, String 
+    public BouteilleBiere(String unNom, double unDegre, String 
         
             uneBrasserie) {
             nom = unNom;
